# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация",
        "description": "Документация для разработчкиков",
        "version": "1.0"
    },
    "paths": {
        "/posts/{id}": {
            "get": {
                "operationId": "c8bd68bc05422c879bc2dd429ae5782b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Title"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Description post"
                                        },
                                        "short_description": {
                                            "type": "string",
                                            "example": "Short description post"
                                        },
                                        "deleted_at": {
                                            "type": "null",
                                            "example": "null"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-10-07T15:57:01.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2022-10-07T15:57:01.000000Z"
                                        },
                                        "cover": {
                                            "type": "string",
                                            "example": "https://via.placeholder.com/1000"
                                        },
                                        "miniature": {
                                            "type": "string",
                                            "example": "https://via.placeholder.com/400"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "05f7903ba196a219ad3b6622c3b2fed4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}